take jetbrains common functionality and put it on its own utils file
