+++ src/editors/editor.ts
@@ -1,7 +1,5 @@
 import fs from "fs";
 import util from "util";
-import path from "path";
-import request from "request";
 
 const stat = util.promisify(fs.stat);
 
@@ -55,45 +53,6 @@ export default abstract class Editor implements EditorInterface {
     } catch (err) {
       // console.error(err);
       return false;
-    }
-  }
-
-  public async installJetbrainsPlugin(pluginsDirectory: string): Promise<void> {
-    let temp = path.join(pluginsDirectory);
-
-    // Create the temp folder first if this does not exists yet
-    fs.mkdirSync(temp, { recursive: true });
-
-    temp = path.join(temp, "WakaTime.jar");
-
-    const file = fs.createWriteStream(temp);
-
-    await new Promise((resolve, reject) => {
-      request({
-        uri: "https://plugins.jetbrains.com/files/7425/51419/WakaTime.jar"
-      })
-        .pipe(file)
-        .on("finish", async () => {
-          resolve();
-        })
-        .on("error", (err: any) => {
-          console.error(err);
-          reject(err);
-        });
-    }).catch(err => {
-      console.error(err);
-    });
-  }
-
-  public async unInstallJetbrainsPlugin(
-    pluginsDirectory: string
-  ): Promise<void> {
-    try {
-      fs.unlinkSync(`${pluginsDirectory}/WakaTime.jar`);
-      return Promise.resolve();
-    } catch (err) {
-      console.error(err);
-      return Promise.reject();
     }
   }
